requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
click==8.1.4
cryptography==41.0.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.98.0
fastapi-users==12.0.0
fastapi-users-db-sqlalchemy==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
makefun==1.15.1
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.1.2
PyJWT==2.7.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-env==0.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.22.0
